+++ components/Widget.tsx
@@ -3,6 +3,7 @@ import Giscus from '../components/Giscus';
 import { AuthContext, ConfigContext, getLoginUrl } from '../lib/context';
 import { emitData } from '../lib/messages';
 import { IErrorMessage, IResizeHeightMessage } from '../lib/types/giscus';
+import { cleanAnchor } from '../lib/utils';
 import { createDiscussion } from '../services/giscus/createDiscussion';
 import { getToken } from '../services/giscus/token';
 
@@ -20,7 +21,7 @@ export default function Widget({ origin, session }: IWidgetProps) {
       repositoryId: repoId,
       categoryId,
       title: term,
-      body: `# ${term}\n\n${description || ''}\n\n${origin}`,
+      body: `# ${term}\n\n${description || ''}\n\n${cleanAnchor(origin)}`,
     });
 
   const handleError = useCallback(